Change method used to retrieve subtype item name
This will help remove phantom microblocks from mods that will allow a
block to be made with a certain meta, but have no localization or sue
for said block.  Mostly this helps to clean up the NEI item listing.
Now the only mod that I've encountered so far that still makes some bad
listings is Forestry;, it will make microblocks for "wood.24 Wood" for
example.