Fix pipes sometimes dispatching an extra stack
To recreate the bug: place two chests 10 blocks apart, connected by pipes.
Place an extractor chip in the interface pipe that was placed first (the order
is important), and a responder chip in the other interface pipe. Both chips
should be configured to blacklist nothing.
Fill all slots of the responder's chest. Have at least two stacks of stuff in
the extractor's chest. Now, remove a stack from the responder's chest.
If you did it right, you should see a stack being sent to the responder's chest
to fill up the single empty slot... then, once it has arrived, a second stack
gets sent, even though the chest is already full! This results in a stack of
overflow in the pipe.

The bug was caused by the PayloadArrivedEvent being sent too early. It was sent
as soon as the item enters the final pipe, even though it hasn't entered the
target chest at that point yet. Since the event removes the stack from the
transitQueue, if the extractor pipe ticks shortly after, what it will see is
that the target chest has an empty slot (since the item still hasn't entered it
yet), but there isn't any item in the transit queue, so it thinks the stack
wasn't sent yet. So it sends it again!

Delaying the PayloadArrivedEvent until the item has actually entered the chest
fixes this bug.

The reason the bug didn't happen all the time was that it requires both of the
following:
1. Within one tick, the extractor pipe needs to tick before the responder pipe
2. The extractor pipe's operation delay must reach 0 before the item enters the
   chest, but after the PayloadArrivedEvent has fired