Only fire PayloadArrivedEvent if passing to inventory is successful
passToInventory is called (and returns false) even when passing a stack into a
pipe, so firing PayloadArrivedEvents at that point was a mistake.

This fixes extra stacks being pushed into the network when a stack bounces back
from a destination that's the wrong one for that stack.
To recreate this, create a T-shaped pipe network with 1 extractor and 2
responders, but no junction pipes.